Starting to refactor code to switch to React Router
