+++ client/components/activity-detail.jsx
@@ -26,8 +26,6 @@ class ActivityDetail extends React.Component {
         <button
           className="spon-button rounded text-white mt-0"
           onClick={() => {
-            const activityId = activity.activityId;
-            this.props.reserve({ activityId });
             this.props.history.push('/activity-details/confirmed');
           }}>
         Confirm
@@ -53,7 +51,7 @@ class ActivityDetail extends React.Component {
         {this.state.showModal ? (
           <CancelModal
             cancel={this.props.cancel}
-            activityId={this.props.activity.activityId}
+            activityId={activity.activityId}
           />
         ) : null}
         <button
