+++ client/components/cancel-modal.jsx
@@ -1,23 +1,58 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 class CancelModal extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+    this.reserveConfirmAndCancel = this.reserveConfirmAndCancel.bind(this);
+  }
+
+  componentDidMount() {
+    this.reserveConfirmAndCancel();
+  }
+
+  reserveConfirmAndCancel({ activityId }) {
+    const config = {
+      method: 'POST',
+      body: JSON.stringify({ activityId }),
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    };
+    fetch('/api/reservations', config).then(response => response.json());
+  }
 
   render() {
     return (
       <div className="confirm-cancel-modal container">
         <div className="inner-modal text-center p-3 rounded">
-          <h3 className="mb-3">Are you sure you want to cancel your reservation?</h3>
-          <h5>Cancelling costs you <span className="font-color">50</span> Spontaneity Points</h5>
+          <h3 className="mb-3">
+            Are you sure you want to cancel your reservation?
+          </h3>
+          <h5>
+            Cancelling costs you <span className="font-color">50</span>{' '}
+            Spontaneity Points
+          </h5>
           <div className="calc-button-50">
-            <button
-              className="spon-button-cancel rounded"
-              onClick={() => {
-                const activityId = this.props.activityId;
-                this.props.cancel({ activityId });
-                this.props.closeModal();
-                this.props.history.push('/');
-              }}>Confirm</button>
-            <button className="spon-button text-white rounded" onClick={this.props.closeModal}>Back</button>
+            <Link to="/">
+              <button
+                className="spon-button-cancel rounded"
+                onClick={() => {
+                  const activityId = this.props.activityId;
+                  this.reserveConfirmAndCancel({ activityId });
+                }}
+              >
+                Confirm
+              </button>
+            </Link>
+            <Link to="/">
+              <button
+                className="spon-button text-white rounded"
+              >
+                Back
+              </button>
+            </Link>
           </div>
         </div>
       </div>
